# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BDE Farm Trac API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/mine:
    get:
      operationId: getMyTenants
      tags: [tenants]
      summary: List tenants the current user belongs to
      responses:
        "200":
          description: List of user tenants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTenantsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenants:
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantBody"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantEnvelope"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenants/current:
    get:
      operationId: getCurrentTenant
      tags: [tenants]
      summary: Get the current tenant
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      responses:
        "200":
          description: Current tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantEnvelope"
    put:
      operationId: updateCurrentTenant
      tags: [tenants]
      summary: Update the current tenant
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantBody"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantEnvelope"
  /tenants/current/farms:
    get:
      operationId: listFarms
      tags: [tenants]
      summary: List farms for the current tenant
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      responses:
        "200":
          description: Farm list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmListResponse"
    post:
      operationId: createFarm
      tags: [tenants]
      summary: Create a farm
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFarmBody"
      responses:
        "201":
          description: Farm created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmEnvelope"
  /tenants/current/farms/{farmId}:
    put:
      operationId: updateFarm
      tags: [tenants]
      summary: Update a farm
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
        - name: farmId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFarmBody"
      responses:
        "200":
          description: Updated farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmEnvelope"
  /tenants/current/invitations:
    get:
      operationId: listInvitations
      tags: [tenants]
      summary: List invitations for the current tenant
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      responses:
        "200":
          description: Invitation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationListResponse"
    post:
      operationId: createInvitation
      tags: [tenants]
      summary: Invite a user to the tenant
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationBody"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationEnvelope"
  /tenants/current/staff-assignments:
    get:
      operationId: listStaffAssignments
      tags: [tenants]
      summary: List staff farm assignments
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      responses:
        "200":
          description: Assignment list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentListResponse"
    post:
      operationId: createStaffAssignment
      tags: [tenants]
      summary: Assign staff to a farm
      parameters:
        - $ref: "#/components/parameters/TenantSlugHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffAssignmentBody"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentEnvelope"
...
components:
  parameters:
...
    TenantSlugHeader:
      name: x-tenant-slug
      in: header
      required: true
      description: Tenant slug to scope the request
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        isActive:
          type: boolean
        stripeCustomerId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, contactEmail, isActive, createdAt]
    TenantEnvelope:
      type: object
      properties:
        tenant:
          $ref: "#/components/schemas/Tenant"
      required: [tenant]
    CreateTenantBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
        address:
          type: string
      required: [name, slug, contactEmail]
    UpdateTenantBody:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
    UserTenantsResponse:
      type: object
      properties:
        tenants:
          type: array
          items:
            type: object
            properties:
              tenantId:
                type: integer
              roleId:
                type: integer
              isSuperAdmin:
                type: boolean
              tenantName:
                type: string
              tenantSlug:
                type: string
            required: [tenantId, roleId, isSuperAdmin, tenantName, tenantSlug]
      required: [tenants]
    Farm:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        postcode:
          type: ["string", "null"]
        cphNumber:
          type: ["string", "null"]
        gridReference:
          type: ["string", "null"]
        totalAcreage:
          type: ["integer", "null"]
        sectorArable:
          type: boolean
        sectorBeef:
          type: boolean
        sectorDairy:
          type: boolean
        sectorPigs:
          type: boolean
        sectorPoultry:
          type: boolean
        sectorHorticulture:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, name, isActive]
    FarmEnvelope:
      type: object
      properties:
        farm:
          $ref: "#/components/schemas/Farm"
      required: [farm]
    FarmListResponse:
      type: object
      properties:
        farms:
          type: array
          items:
            $ref: "#/components/schemas/Farm"
      required: [farms]
    CreateFarmBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        postcode:
          type: string
        cphNumber:
          type: string
        gridReference:
          type: string
        totalAcreage:
          type: integer
        sectorArable:
          type: boolean
        sectorBeef:
          type: boolean
        sectorDairy:
          type: boolean
        sectorPigs:
          type: boolean
        sectorPoultry:
          type: boolean
        sectorHorticulture:
          type: boolean
      required: [name]
    InvitationListResponse:
      type: object
      properties:
        invitations:
          type: array
          items:
            type: object
      required: [invitations]
    CreateInvitationBody:
      type: object
      properties:
        email:
          type: string
          format: email
        roleId:
          type: integer
      required: [email, roleId]
    InvitationEnvelope:
      type: object
      properties:
        invitation:
          type: object
      required: [invitation]
    AssignmentListResponse:
      type: object
      properties:
        assignments:
          type: array
          items:
            type: object
      required: [assignments]
    CreateStaffAssignmentBody:
      type: object
      properties:
        userId:
          type: string
        farmId:
          type: integer
        roleId:
          type: integer
      required: [userId, farmId]
    AssignmentEnvelope:
      type: object
      properties:
        assignment:
          type: object
      required: [assignment]
...